+++ lib/dropsync.rb
@@ -1,0 +1,5 @@
+require 'dropsync/dropsync'
+require 'dropsync/version'
+require 'dropsync/mecha'
+
+require 'dropbox_sdk'
