+++ lib/dropsync/version.rb
@@ -1,0 +1,3 @@
+module DropSync
+  VERSION = "1.0.0.pre"
+end
